ref: Move util file in utils instead of fragments
